- Completely redo this module > jk
Actually no, I just made it a bit more pretty

+++ userbot/plugins/corona.py
@@ -1,18 +1,39 @@
 import requests
+from prettytable import PrettyTable
 from time import sleep
 from userbot import UserBot
 from pyrogram import Filters, Message
 from userbot.plugins.help import add_command_help
-
+import datetime
 
 @UserBot.on_message(Filters.command(["c"], ".") & Filters.me)
 async def corona_all(bot: UserBot, message: Message):
     try:
-        r = requests.get("https://corona.lmao.ninja/v2/all").json()
-        all_cases = f"<b>Global Totals</b>\nCases: {r['cases']:,}\nDeaths: {r['deaths']:,}\nRecovered: {r['recovered']:,}\nActive: {r['active']:,}"
-        await message.edit(all_cases)
-    except:
+        r = requests.get("https://corona.lmao.ninja/v2/all?yesterday=true").json()
+        #all_cases = f"<b>Global Totals</b>\nCases: {r['cases']:,}\nDeaths: {r['deaths']:,}\nRecovered: {r['recovered']:,}\nActive: {r['active']:,}"
+        last_updated = datetime.datetime.fromtimestamp(r['updated']/1000).strftime("%Y-%m-%d %I:%M:%S")
+        
+        ac = PrettyTable()
+        ac.header = False
+        ac.title = "Global Statistics"
+        
+        ac.add_row(["Cases", f"{r['cases']:,}"])
+        ac.add_row(["Cases Today", f"{r['todayCases']:,}"])
+        ac.add_row(["Deaths", f"{r['deaths']:,}"])
+        ac.add_row(["Deaths Today", f"{r['todayDeaths']:,}"])
+        ac.add_row(["Recovered", f"{r['recovered']:,}"])
+        ac.add_row(["Active", f"{r['active']:,}"])
+        ac.add_row(["Critical", f"{r['critical']:,}"])
+        ac.add_row(["Cases/Million", f"{r['casesPerOneMillion']:,}"])
+        ac.add_row(["Deaths/Million", f"{r['deathsPerOneMillion']:,}"])
+        ac.add_row(["Tests", f"{r['tests']:,}"])
+        ac.add_row(["Tests/Million", f"{r['testsPerOneMillion']:,}"])
+        #ac.add_row(["Affected Countries", f"{r['affectedCountries']:,}"])
+        ac.align = "l"
+        await message.edit(f"```{str(ac)}```\nLast updated on: {last_updated}")
+    except Exception as e:
         await message.edit("```The corona API could not be reached```")
+        print(e)
         sleep(3)
         await message.delete()
 
@@ -30,15 +51,33 @@ async def corona_search(bot: UserBot, message: Message):
     country = cmd[1]
     await message.edit(f"```Getting Corona statistics for {country}```")
 
-    try:
-        r = requests.get(f"https://corona.lmao.ninja/v2/countries/{country}").json()
-    except:
-        await message.edit("```The corona API could not be reached```")
+    
+    r = requests.get(f"https://corona.lmao.ninja/v2/countries/{country}").json()
+    if "cases" not in r:
+        await message.edit("```The country could not be found!```")
         sleep(3)
         await message.delete()
     else:
-        country_cases = f"<b>Cases for {r['country']}</b>\nCases: {r['cases']:,}\nCases Today: {r['todayCases']:,}\nDeaths: {r['deaths']:,}\nDeaths Today: {r['todayDeaths']:,}\nRecovered: {r['recovered']:,}\nActive: {r['active']:,}\nCritical: {r['critical']:,}\nCases/Mil: {r['casesPerOneMillion']}\nDeaths/Mil: {r['deathsPerOneMillion']}"
-        await message.edit(country_cases)
+        #country_cases = f"<b>Cases for {r['country']}</b>\nCases: {r['cases']:,}\nCases Today: {r['todayCases']:,}\nDeaths: {r['deaths']:,}\nDeaths Today: {r['todayDeaths']:,}\nRecovered: {r['recovered']:,}\nActive: {r['active']:,}\nCritical: {r['critical']:,}\nCases/Mil: {r['casesPerOneMillion']}\nDeaths/Mil: {r['deathsPerOneMillion']}"
+        last_updated = datetime.datetime.fromtimestamp(r['updated']/1000).strftime("%Y-%m-%d %I:%M:%S")
+        
+        cc = PrettyTable()
+        cc.header = False
+        country = r['countryInfo']['iso3'] if len(r['country']) > 12 else r['country']
+        cc.title = f"Corona Cases in {country}"
+        cc.add_row(["Cases", f"{r['cases']:,}"])
+        cc.add_row(["Cases Today", f"{r['todayCases']:,}"])
+        cc.add_row(["Deaths", f"{r['deaths']:,}"])
+        cc.add_row(["Deaths Today", f"{r['todayDeaths']:,}"])
+        cc.add_row(["Recovered", f"{r['recovered']:,}"])
+        cc.add_row(["Active", f"{r['active']:,}"])
+        cc.add_row(["Critical", f"{r['critical']:,}"])
+        cc.add_row(["Cases/Million", f"{r['casesPerOneMillion']:,}"])
+        cc.add_row(["Deaths/Million", f"{r['deathsPerOneMillion']:,}"])
+        cc.add_row(["Tests", f"{r['tests']:,}"])
+        cc.add_row(["Tests/Million", f"{r['testsPerOneMillion']:,}"])
+        cc.align = "l"
+        await message.edit(f"```{str(cc)}```\nLast updated on: {last_updated}")
 
 
 add_command_help(
